[Support] Move Directory class to Directory namespace.
