+++ tests/Unit/Http/Middleware/Cache/CacheResponseMiddlewareTest.php
@@ -23,7 +23,7 @@ use Valkyrja\Http\Message\Response\EmptyResponse;
 use Valkyrja\Http\Middleware\Cache\CacheResponseMiddleware;
 use Valkyrja\Http\Middleware\Handler\RequestReceivedHandler;
 use Valkyrja\Http\Middleware\Handler\TerminatedHandler;
-use Valkyrja\Support\Directory;
+use Valkyrja\Support\Directory\Directory;
 use Valkyrja\Support\Time\Time;
 use Valkyrja\Tests\Unit\TestCase;
